+++ CHANGELOG.md
@@ -8,7 +8,7 @@ FoundryWorldTools first release. Only works on Unix
   * Updated complete example to include seperate commands for Windows
 * Added ability to rename / copy world directories
 
-# April 25, 2021
+# April 26, 2021
 
 This release is a complete rewrite of the utility with more robust path
 support. Better config management, logging, and new commands and options have 
@@ -22,3 +22,9 @@ been added.
 * Added the pull command
 * Added --replace option to rewriteall
 * Added --lower option to rewriteall
+
+# May 26
+
+* BUG FIX: Also update paths in the project's manafest JSON file
+* BUG FIX: Properly handle --dataDir option
+* FEATURE: Added --exclude-dir option on dedupe command